+++ cmd/login/config.go
@@ -1,0 +1,9 @@
+package login
+
+const (
+	ClientID     = "client-id"
+	ClientSecret = "client-secret"
+	Issuer       = "idp-issuer-url"
+	IDToken      = "id-token"
+	RefreshToken = "refresh-token"
+)