+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+### Changed
+
+- Make the `login` command validate the current context before considering it good to go.
+
 ## [1.20.0] - 2021-01-18
 
 ### Added